Extract out all the fields
