+++ frontend/src/js/external-forms/form/Field.tsx
@@ -1,170 +1,27 @@
-import styled from "@emotion/styled";
-import { memo, ReactNode } from "react";
-import {
-  Control,
-  ControllerRenderProps,
-  useController,
-  UseFormRegister,
-  UseFormSetValue,
-} from "react-hook-form";
-import { useTranslation } from "react-i18next";
+import { memo } from "react";
+import { Control, UseFormRegister, UseFormSetValue } from "react-hook-form";
 
 import type { SelectOptionT } from "../../api/types";
-import type { DateStringMinMax } from "../../common/helpers/dateHelper";
-import { exists } from "../../common/helpers/exists";
 import { useDatasetId } from "../../dataset/selectors";
 import type { Language } from "../../localization/useActiveLang";
-import { nodeIsInvalid } from "../../model/node";
-import type { DragItemQuery } from "../../standard-query-editor/types";
-import InputCheckbox from "../../ui-components/InputCheckbox";
-import InputDateRange from "../../ui-components/InputDateRange";
-import InputPlain from "../../ui-components/InputPlain/InputPlain";
-import InputSelect from "../../ui-components/InputSelect/InputSelect";
-import { InputTextarea } from "../../ui-components/InputTextarea/InputTextarea";
-import ToggleButton from "../../ui-components/ToggleButton";
-import type { Field as FieldT, GeneralField, Tabs } from "../config-types";
+import type { GeneralField } from "../config-types";
 import { Description } from "../form-components/Description";
-import {
-  getHeadlineFieldAs,
-  Headline,
-  HeadlineIndex,
-} from "../form-components/Headline";
-import FormConceptGroup from "../form-concept-group/FormConceptGroup";
-import type { FormConceptGroupT } from "../form-concept-group/formConceptGroupState";
-import FormQueryDropzone from "../form-query-dropzone/FormQueryDropzone";
-import FormTabNavigation from "../form-tab-navigation/FormTabNavigation";
-import { getFieldKey, getInitialValue, isFormField } from "../helper";
-import { getErrorForField } from "../validators";
+import { getInitialValue, isFormField } from "../helper";
 
 import type { DynamicFormValues } from "./Form";
-
-// TODO: REFINE COLORS
-// const useColorByField = (fieldType: FormField["type"]) => {
-//   const theme = useTheme();
-
-//   const COLOR_BY_FIELD_TYPE: Record<FormField["type"], string> = useMemo(
-//     () => ({
-//       STRING: theme.col.palette[8],
-//       DATE_RANGE: theme.col.palette[0],
-//       NUMBER: theme.col.palette[1],
-//       CONCEPT_LIST: theme.col.palette[2],
-//       SELECT: theme.col.palette[3],
-//       DATASET_SELECT: theme.col.palette[4],
-//       CHECKBOX: theme.col.palette[7],
-//       RESULT_GROUP: theme.col.palette[5],
-//       TABS: theme.col.palette[9],
-//     }),
-//     [theme],
-//   );
-
-//   return COLOR_BY_FIELD_TYPE[fieldType];
-// };
-
-type Props<T> = T & {
-  children: (props: ControllerRenderProps<DynamicFormValues>) => ReactNode;
-  control: Control<DynamicFormValues>;
-  formField: FieldT | Tabs;
-  defaultValue?: unknown;
-  noContainer?: boolean;
-  noLabel?: boolean;
-};
-const FieldContainer = styled("div")<{
-  noLabel?: boolean;
-  hasError?: boolean;
-  red?: boolean;
-}>`
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-  padding: ${({ noLabel }) => (noLabel ? "7px 10px" : "2px 10px 7px")};
-  background-color: white;
-  border-radius: ${({ theme }) => theme.borderRadius};
-  border: 1px solid
-    ${({ theme, hasError, red }) =>
-      hasError
-        ? red
-          ? theme.col.red
-          : theme.col.blueGrayDark
-        : theme.col.grayLight};
-`;
-
-const ErrorContainer = styled("div")<{ red?: boolean }>`
-  color: ${({ theme, red }) => (red ? theme.col.red : theme.col.blueGrayDark)};
-  font-weight: 700;
-  font-size: ${({ theme }) => theme.font.sm};
-`;
-
-const ConnectedField = <T extends object>({
-  children,
-  control,
-  formField,
-  defaultValue,
-  noContainer,
-  noLabel,
-  ...props
-}: Props<T>) => {
-  const { t } = useTranslation();
-  const { field, fieldState } = useController<DynamicFormValues>({
-    name: formField.name,
-    defaultValue,
-    control,
-    rules: {
-      validate: (value) => getErrorForField(t, formField, value) || true,
-    },
-    shouldUnregister: false,
-  });
-
-  // TODO: REFINE COLORS
-  // const color = useColorByField(formField.type);
-
-  const requiredMsg = t("externalForms.formValidation.isRequired");
-  const isRedError = fieldState.error?.message !== requiredMsg;
-
-  return noContainer ? (
-    <div>{children({ ...field, ...props })}</div>
-  ) : (
-    <FieldContainer
-      noLabel={noLabel}
-      hasError={exists(fieldState.error)}
-      red={isRedError}
-    >
-      {children({ ...field, ...props })}
-      <ErrorContainer red={isRedError}>
-        {fieldState.error?.message}
-      </ErrorContainer>
-    </FieldContainer>
-  );
-};
-
-const SxToggleButton = styled(ToggleButton)`
-  margin-bottom: 5px;
-`;
-
-const Spacer = styled("div")`
-  height: 14px;
-`;
-
-const Group = styled("div")`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-`;
-
-const NestedFields = styled("div")`
-  display: flex;
-  flex-direction: column;
-  gap: 7px;
-  padding: 12px 10px 12px;
-  background-color: ${({ theme }) => theme.col.bg};
-  border: 1px solid ${({ theme }) => theme.col.gray};
-  border-radius: ${({ theme }) => theme.borderRadius};
-`;
-
-const setValueConfig = {
-  shouldValidate: true,
-  shouldDirty: true,
-  shouldTouch: true,
-};
+import { CheckboxField } from "./fields/CheckboxField";
+import { ConceptListField } from "./fields/ConceptListField";
+import { DatasetSelectField } from "./fields/DatasetSelectField";
+import { DateRangeField } from "./fields/DateRangeField";
+import { DisclosureField } from "./fields/DisclosureField";
+import { GroupField } from "./fields/GroupField";
+import { HeadlineField } from "./fields/HeadlineField";
+import { NumberField } from "./fields/NumberField";
+import { ResultGroupField } from "./fields/ResultGroupField";
+import { SelectField } from "./fields/SelectField";
+import { StringField } from "./fields/StringField";
+import { TabsField } from "./fields/TabsField";
+import { TextAreaField } from "./fields/TextAreaField";
 
 const Field = ({
   field,
@@ -180,9 +37,7 @@ const Field = ({
   control: Control<DynamicFormValues>;
 }) => {
   const datasetId = useDatasetId();
-  const { formType, h1Index, locale, availableDatasets, setValue, control } =
-    commonProps;
-  const { t } = useTranslation();
+  const { locale, availableDatasets } = commonProps;
 
   const defaultValue =
     isFormField(field) && field.type !== "GROUP"
@@ -195,12 +50,7 @@ const Field = ({
 
   switch (field.type) {
     case "HEADLINE":
-      return (
-        <Headline as={getHeadlineFieldAs(field)} size={field.style?.size}>
-          {exists(h1Index) && <HeadlineIndex>{h1Index + 1}</HeadlineIndex>}
-          {field.label[locale]}
-        </Headline>
-      );
+      return <HeadlineField field={field} commonProps={commonProps} />;
     case "DESCRIPTION":
       return (
         <Description
@@ -209,353 +59,93 @@ const Field = ({
       );
     case "STRING":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <StringField
+          field={field}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => (
-            <InputPlain
-              ref={ref}
-              inputType="text"
-              label={field.label[locale] || ""}
-              placeholder={
-                (field.placeholder && field.placeholder[locale]) || ""
-              }
-              fullWidth={field.style ? field.style.fullWidth : false}
-              value={fieldProps.value as string}
-              onChange={(value) => setValue(field.name, value, setValueConfig)}
-              tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-            />
-          )}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "TEXTAREA":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <TextAreaField
+          field={field}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => (
-            <InputTextarea
-              ref={ref}
-              label={field.label[locale] || ""}
-              placeholder={
-                (field.placeholder && field.placeholder[locale]) || ""
-              }
-              rows={field.style?.rows ?? 4}
-              value={fieldProps.value as string}
-              onChange={(value) => {
-                setValue(field.name, value, setValueConfig);
-              }}
-              tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-            />
-          )}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "NUMBER":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <NumberField
+          field={field}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => (
-            <InputPlain
-              ref={ref}
-              inputType="number"
-              label={field.label[locale] || ""}
-              placeholder={
-                (field.placeholder && field.placeholder[locale]) || ""
-              }
-              value={fieldProps.value as number | null}
-              onChange={(value) => setValue(field.name, value, setValueConfig)}
-              inputProps={{
-                step: field.step || "1",
-                pattern: field.pattern,
-                min: field.min,
-                max: field.max,
-              }}
-              tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-            />
-          )}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "DATE_RANGE":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <DateRangeField
+          field={field}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => {
-            return (
-              <InputDateRange
-                inline={true}
-                label={field.label[locale]}
-                tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-                value={fieldProps.value as DateStringMinMax}
-                onChange={(value) =>
-                  setValue(field.name, value, setValueConfig)
-                }
-              />
-            );
-          }}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "RESULT_GROUP":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <ResultGroupField
+          field={field}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => (
-            <FormQueryDropzone
-              label={field.label[locale] || ""}
-              dropzoneText={field.dropzoneLabel[locale] || ""}
-              tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-              value={fieldProps.value as DragItemQuery}
-              onChange={(value) => setValue(field.name, value, setValueConfig)}
-            />
-          )}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "CHECKBOX":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <CheckboxField
+          field={field}
           defaultValue={defaultValue}
-          noLabel
-        >
-          {({ ref, ...fieldProps }) => (
-            <InputCheckbox
-              value={fieldProps.value as boolean}
-              onChange={(value) => setValue(field.name, value, setValueConfig)}
-              label={field.label[locale] || ""}
-              infoTooltip={field.tooltip ? field.tooltip[locale] : undefined}
-            />
-          )}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "SELECT":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
+        <SelectField
+          field={field}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => (
-            <InputSelect
-              label={field.label[locale]}
-              options={field.options.map((option) => ({
-                label: option.label[locale] || "",
-                value: option.value,
-              }))}
-              tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-              value={fieldProps.value as SelectOptionT | null}
-              onChange={(value) => setValue(field.name, value, setValueConfig)}
-            />
-          )}
-        </ConnectedField>
+          commonProps={commonProps}
+        />
       );
     case "DATASET_SELECT":
       return (
-        <ConnectedField
-          formField={field}
-          control={control}
-          defaultValue={
-            availableDatasets.length > 0
-              ? availableDatasets.find((opt) => opt.value === datasetId) ||
-                availableDatasets[0]
-              : null
-          }
-        >
-          {({ ref, ...fieldProps }) => {
-            return (
-              <InputSelect
-                label={field.label[locale]}
-                options={availableDatasets}
-                tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-                value={fieldProps.value as SelectOptionT | null}
-                onChange={(value) =>
-                  setValue(field.name, value, setValueConfig)
-                }
-              />
-            );
-          }}
-        </ConnectedField>
+        <DatasetSelectField
+          field={field}
+          commonProps={commonProps}
+          datasetId={datasetId}
+        />
       );
     case "DISCLOSURE":
       return (
-        <details>
-          <summary>
-            <Headline>{field.label[locale]}</Headline>
-          </summary>
-          {field.fields.map((f, i) => {
-            const key = getFieldKey(formType, f, i);
-
-            return <Field key={key} field={f} {...commonProps} />;
-          })}
-        </details>
+        <DisclosureField
+          field={field}
+          commonProps={commonProps}
+          key={field.name}
+        />
       );
     case "GROUP":
-      return (
-        <>
-          {field.label && <Headline>{field.label[locale]}</Headline>}
-          {field.description && (
-            <Description>{field.description[locale]}</Description>
-          )}
-          <Group
-            style={{
-              display: (field.style && field.style.display) || "flex",
-              gap: field.style?.display === "grid" ? "7px 12px" : "7px 8px",
-              gridTemplateColumns: `repeat(${
-                field.style?.gridColumns || 1
-              }, 1fr)`,
-            }}
-          >
-            {field.fields.map((f, i) => {
-              const key = getFieldKey(formType, f, i);
-
-              return <Field key={key} field={f} {...commonProps} />;
-            })}
-          </Group>
-        </>
-      );
+      return <GroupField field={field} commonProps={commonProps} />;
     case "TABS":
       return (
-        <ConnectedField
-          control={control}
-          formField={field}
+        <TabsField
+          field={field}
+          commonProps={commonProps}
           defaultValue={defaultValue}
-          noContainer
-        >
-          {({ ref, ...fieldProps }) => {
-            const tabToShow = field.tabs.find(
-              (tab) => tab.name === fieldProps.value,
-            );
-
-            return (
-              <>
-                <FormTabNavigation
-                  selectedTab={fieldProps.value as string}
-                  onSelectTab={(tab) =>
-                    setValue(field.name, tab, setValueConfig)
-                  }
-                  options={field.tabs.map((tab) => ({
-                    label: () => tab.title[locale] || "",
-                    value: tab.name,
-                    tooltip: tab.tooltip ? tab.tooltip[locale] : undefined,
-                  }))}
-                />
-                {tabToShow && tabToShow.fields.length > 0 ? (
-                  <NestedFields>
-                    {tabToShow.fields.map((f, i) => {
-                      const key = getFieldKey(formType, f, i);
-
-                      return <Field key={key} field={f} {...commonProps} />;
-                    })}
-                  </NestedFields>
-                ) : (
-                  <Spacer />
-                )}
-              </>
-            );
-          }}
-        </ConnectedField>
+        />
       );
     case "CONCEPT_LIST":
       return (
-        <ConnectedField
-          control={control}
-          formField={field}
+        <ConceptListField
+          field={field}
+          commonProps={commonProps}
           defaultValue={defaultValue}
-        >
-          {({ ref, ...fieldProps }) => (
-            <FormConceptGroup
-              fieldName={field.name}
-              value={fieldProps.value as FormConceptGroupT[]}
-              onChange={(value) => setValue(field.name, value, setValueConfig)}
-              label={field.label[locale] || ""}
-              tooltip={field.tooltip ? field.tooltip[locale] : undefined}
-              conceptDropzoneText={
-                field.conceptDropzoneLabel
-                  ? field.conceptDropzoneLabel[locale] || ""
-                  : t("externalForms.default.conceptDropzoneLabel")
-              }
-              attributeDropzoneText={
-                field.conceptColumnDropzoneLabel
-                  ? field.conceptColumnDropzoneLabel[locale] || ""
-                  : t("externalForms.default.conceptDropzoneLabel")
-              }
-              formType={formType}
-              disallowMultipleColumns={!field.isTwoDimensional}
-              isSingle={field.isSingle}
-              blocklistedTables={field.blocklistedConnectors}
-              allowlistedTables={field.allowlistedConnectors}
-              blocklistedSelects={field.blocklistedSelects}
-              allowlistedSelects={field.allowlistedSelects}
-              defaults={field.defaults}
-              isValidConcept={(item) =>
-                !nodeIsInvalid(
-                  item,
-                  field.blocklistedConceptIds,
-                  field.allowlistedConceptIds,
-                )
-              }
-              // What follows is VERY custom
-              // Concept Group supports rendering a prefix field
-              // That's specifically required by one of the forms: "PSM Form"
-              // So the following looks like it wants to be generic,
-              // but it's really implemented for one field
-              newValue={
-                field.rowPrefixField
-                  ? {
-                      concepts: [],
-                      connector: "OR",
-                      [field.rowPrefixField.name]:
-                        field.rowPrefixField.defaultValue,
-                    }
-                  : { concepts: [], connector: "OR" }
-              }
-              rowPrefixFieldname={field.rowPrefixField?.name}
-              renderRowPrefix={
-                exists(field.rowPrefixField)
-                  ? ({ value: fieldValue, onChange, row, i }) => (
-                      <SxToggleButton
-                        options={field.rowPrefixField!.options.map(
-                          (option) => ({
-                            label: option.label[locale] || "",
-                            value: option.value,
-                          }),
-                        )}
-                        value={
-                          /* Because we're essentially adding an extra dynamic field to FormConceptGroupT
-                            with the key `field.rowPrefixField.name` */
-                          (row as unknown as Record<string, string>)[
-                            field.rowPrefixField!.name
-                          ]
-                        }
-                        onChange={(value) =>
-                          onChange([
-                            ...fieldValue.slice(0, i),
-                            {
-                              ...fieldValue[i],
-                              [field.rowPrefixField!.name]: value,
-                            },
-                            ...fieldValue.slice(i + 1),
-                          ])
-                        }
-                      />
-                    )
-                  : undefined
-              }
-            />
-          )}
-        </ConnectedField>
+        />
       );
     default:
       return null;
